Real-time file watching in python
Add a stupid-simple example of tailing a log in real time with
threaded python. The threading-based solution is good for a
simple script where we want to watch one or more logs and then
consume them in workers, but there are issues with resource
cleanup that might prevent it from working in larger programs.

So, also add a multiprocessing-based example where we can rely
in IPC, signals, and process-level cleanup to get similar
functionality but also have the ability to kill all the workers
when we are ready without terminating the main thread.